+++ src/dataProvider/SparqlDataProvider.js
@@ -27,9 +27,10 @@ if (config.queryFolder.substring(config.queryFolder.length - 1) !== "/") {
 
 export default {
   getList: async function getList(queryName, { pagination, sort, filter, meta }) {
+    //console.log("getList in");
     const query = findQueryWithId(queryName);
-    query.limit = pagination.perPage;
-    query.offset = (pagination.page - 1) * pagination.perPage;
+    const limit = pagination.perPage;
+    const offset = (pagination.page - 1) * pagination.perPage;
     query.sort = sort;
 
     handleComunicaContextCreation(query);
@@ -44,6 +45,8 @@ export default {
     }
 
     let results = await executeQuery(query);
+    let totalItems = results.length;
+    results = results.slice(offset, offset + limit);
 
     if (Object.keys(filter).length > 0) {
       results = results.filter((result) => {
@@ -53,9 +56,10 @@ export default {
       });
     }
 
+    //console.log("getList out");
     return {
       data: results,
-      total: query.totalItems
+      total: totalItems
     };
   },
   getOne: async function getOne() {
@@ -115,16 +119,6 @@ async function fetchQuery(query) {
     const parsedQuery = parser.parse(rawText);
     if (!query.variableOntology) {
       query.variableOntology = findPredicates(parsedQuery);
-    }
-    if (parsedQuery.limit !== undefined && query.offset + query.limit > parsedQuery.limit) {
-      parsedQuery.limit = parsedQuery.limit - query.offset;
-    } else {
-      parsedQuery.limit = query.limit;
-    }
-    if (parsedQuery.offset) {
-      parsedQuery.offset += query.offset;
-    } else {
-      parsedQuery.offset = query.offset;
     }
     if (!parsedQuery.order && query.sort && query.sort.field !== "id") {
       const { field, order } = query.sort;
@@ -281,7 +275,6 @@ async function handleQueryExecution(execution, query) {
     const resultType = execution.resultType;
     if (execution.resultType !== "boolean") {
       const metadata = await execution.metadata();
-      query.totalItems = await countQueryResults(query);
       variables = metadata.variables.map((val) => {
         return val.value;
       });
@@ -290,51 +283,6 @@ async function handleQueryExecution(execution, query) {
   } catch (error) {
     throw new HttpError(error.message, 500);
   }
-}
-
-/**
- *
- * @param {object} query - the query which is to be executed and additional information about the query.
- * @returns {number} the actual number of results in the query, if it were executed
- */
-async function countQueryResults(query) {
-  const parser = new Parser();
-  const parsedQuery = parser.parse(query.rawText);
-  const distinctInitial = parsedQuery.distinct;
-  const offsetInitial = parsedQuery.offset;
-  const limitInitial = parsedQuery.limit;
-  parsedQuery.queryType = "SELECT";
-  parsedQuery.distinct = false;
-  parsedQuery.offset = 0;
-  if (parsedQuery.limit) {
-    delete parsedQuery.limit;
-  }
-  parsedQuery.variables = [
-    {
-      expression: {
-        type: "aggregate",
-        aggregation: "count",
-        expression: { termType: "Wildcard", value: "*" },
-        distinct: distinctInitial
-      },
-      variable: { termType: "Variable", value: "totalItems" },
-    },
-  ];
-  const generator = new Generator();
-  const countQuery = generator.stringify(parsedQuery);
-  const bindings = await myEngine.queryBindings(countQuery, {
-    sources: query.comunicaContext.sources,
-    fetch: fetch,
-    httpProxyHandler: proxyHandler,
-  });
-  let totalItems = parseInt((await bindings.toArray())[0].get("totalItems").value);
-  if (offsetInitial) {
-    totalItems -= offsetInitial;
-  }
-  if (limitInitial && totalItems > limitInitial) {
-    totalItems = limitInitial;
-  }
-  return totalItems;
 }
 
 const queryTypeHandlers = {